% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{Generate the Design Matrix for LDA Model}
\usage{
getDesignMatrix(modelLDA, data, scale = FALSE)
}
\arguments{
\item{modelLDA}{A fitted LDA model object containing the terms, variable
indices, variable centers, and scaling factors.}

\item{data}{A data frame containing the predictor variables that are used to
create the design matrix.}

\item{scale}{A logical value indicating whether to scale the design matrix
based on the mean and standard deviation of the variables (default is
\code{FALSE}).}
}
\value{
A design matrix where each row corresponds to an observation and
each column to a predictor variable. If \code{scale = TRUE}, the variables are
centered and scaled based on the means and standard deviations provided in
the LDA model object.
}
\description{
Generate the Design Matrix for LDA Model
}
