% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLDscores}
\alias{getLDscores}
\title{Compute Linear Discriminant Scores}
\usage{
getLDscores(modelLDA, data, nScores = -1)
}
\arguments{
\item{modelLDA}{A fitted LDA model object containing the scaling matrix and
the reference structure for missing data.}

\item{data}{A data frame containing the predictor variables for which to
compute the linear discriminant scores.}

\item{nScores}{An integer specifying the number of discriminant scores to
compute. If \code{-1} (default), all scores are computed.}
}
\value{
A matrix of linear discriminant scores, where rows correspond to
observations and columns correspond to the computed discriminant scores.
If \code{nScores > 0}, only the specified number of scores is returned; otherwise,
all scores are computed and returned.
}
\description{
Compute Linear Discriminant Scores
}
