% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMode}
\alias{getMode}
\title{Calculate the Mode of a Factor Variable with Optional Priors}
\usage{
getMode(v, prior)
}
\arguments{
\item{v}{A factor or vector that can be coerced into a factor. The mode will
be calculated from the levels of this factor.}

\item{prior}{A numeric vector of prior weights for each level of the factor.}
}
\value{
The mode of the factor \code{v} as a character string. If all values are
\code{NA}, the function returns \code{NA}.
}
\description{
Calculate the Mode of a Factor Variable with Optional Priors
}
