% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonConstInd}
\alias{nonConstInd}
\title{Identify Non-Constant Columns in a Data Frame}
\usage{
nonConstInd(data, tol = 1e-08, na.rm = FALSE)
}
\arguments{
\item{data}{A data frame in which columns will be checked for constant
values. Columns can be of any type (numeric, integer, logical, or factor).}

\item{tol}{A numeric tolerance value (default is \code{1e-8}) that applies to
numerical columns.}

\item{na.rm}{A logical value. If \code{FALSE} (default), missing values are
retained during the check.}
}
\value{
An integer vector containing the indices of the non-constant columns
in the data frame. If all columns are constant, an empty vector is
returned.
}
\description{
Identify Non-Constant Columns in a Data Frame
}
