% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSel.R
\name{forwardSel}
\alias{forwardSel}
\title{Title}
\usage{
forwardSel(
  datX,
  response,
  method = "Pillai",
  firstOnly = FALSE,
  alpha = 0.05,
  correction = TRUE,
  fixThreshold = FALSE
)
}
\arguments{
\item{fixThreshold}{}
}
\description{
Title
}
